pyproject.toml
[tool.poetry]
name = "myhouse24"
version = "0.1.0"
description = ""
authors = ["valeriiastartseva <kaleriiastartseva505@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
redis = "^5.2.0"
requests = "^2.32.3"
django-allauth = "^65.1.0"
psycopg2-binary = "^2.9.10"
psycopg = "^3.2.3"
django = "^5.1.2"
celery = "^5.4.0"
django-ajax-datatable = "^4.5.0"
django-debug-toolbar = "^4.4.6"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
